Add configuration option for choosing MESA version
