--- src/amuse/community/mesa/__init__.py
@@ -1,8 +1,0 @@
-"""
-The Astrophysical Multipurpose Software Environment
-
-AMUSE contains a lot of functionality to execute astrophysical experiments.
-Look into the interface packages for the interfacing to physical model simulation software.
-You can also use AMUSE to convert common data formats.
-
-"""
